Fix outlines of FCOs when dragging in model editor under x64
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1819 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/Gme/GMEOLEData.cpp
@@ -66,7 +66,7 @@ void CGMEDataDescriptor::Reset()
 void CGMEDataDescriptor::Serialize(CArchive& ar)
 {
 	if(ar.IsStoring()) {
-		ar << rects.GetCount();
+		ar << (int)rects.GetCount();
 		POSITION pos = rects.GetHeadPosition();
 		while(pos)
 			ar << *(rects.GetNext(pos));
